requirements.txt
fastapi==0.117.1
nltk==3.7
numpy>=1.25.0,<2.0.0
openai==1.108.1
pandas==1.5.3
pydantic==1.10.12
python-dotenv==1.1.1
uvicorn==0.36.0
requests==2.31.0
sentence-transformers==2.6.1
scikit-learn==1.7.2
spacy==3.8.2